requirements.txt
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
numpy==1.14.2
pandas==0.22.0
image==1.5.20
pillow==5.1.0
pytz==2018.4
h5py==2.7.1
Keras==2.1.5
six==1.11.0
tensorflow==1.8.0rc0